Move 'app-kotlin' module to 'com.adobe.marketing.mobile.edge.testapp.kotlin' namespace
