+++ code/app-kotlin/build.gradle.kts
@@ -28,7 +28,7 @@ configure<com.diffplug.gradle.spotless.SpotlessExtension> {
 }
 
 android {
-    namespace = "com.adobe.marketing.tester.app"
+    namespace = "com.adobe.marketing.mobile.edge.testapp.kotlin"
 
     defaultConfig {
         applicationId = "com.adobe.marketing.tester.app"
