--- code/app-kotlin/src/main/java/com/adobe/marketing/tester/app/TestApplication.kt
@@ -1,36 +1,0 @@
-/*
-  Copyright 2023 Adobe. All rights reserved.
-  This file is licensed to you under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License. You may obtain a copy
-  of the License at http://www.apache.org/licenses/LICENSE-2.0
-  Unless required by applicable law or agreed to in writing, software distributed under
-  the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
-  OF ANY KIND, either express or implied. See the License for the specific language
-  governing permissions and limitations under the License.
-*/
-
-package com.adobe.marketing.tester.app
-
-import android.app.Application
-import com.adobe.marketing.mobile.Assurance
-import com.adobe.marketing.mobile.Edge
-import com.adobe.marketing.mobile.LoggingMode
-import com.adobe.marketing.mobile.MobileCore
-import com.adobe.marketing.mobile.edge.consent.Consent
-import com.adobe.marketing.mobile.edge.identity.Identity
-
-class TestApplication : Application() {
-    // TODO: Set up the preferred Environment File ID from your mobile property configured in Data Collection UI
-    private var ENVIRONMENT_FILE_ID: String = ""
-
-    override fun onCreate() {
-        super.onCreate()
-
-        MobileCore.setLogLevel(LoggingMode.VERBOSE)
-        MobileCore.setApplication(this)
-
-        MobileCore.registerExtensions(listOf(Consent.EXTENSION, Identity.EXTENSION, Edge.EXTENSION, Assurance.EXTENSION)) {
-            MobileCore.configureWithAppID(ENVIRONMENT_FILE_ID)
-        }
-    }
-}
